requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community
langchainhub

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

#Groq Integration
langchain-groq

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

#Pydantic for structured output
pydantic
pydantic[email]

#document-load
pypdf

#vector store 
tiktoken
faiss-cpu

# Streamlit - for UI
streamlit

# YouTube API and content extraction
youtube-transcript-api
youtube-search-python

#Article Scraping
newspaper3k 
